+++ test/unit/app/exception/unhandledRejection.test.js
@@ -2,7 +2,7 @@ describe('Unhandled Rejection Test', () => {
   test('Should handle rejection', async () => {
     const mockExit = jest.spyOn(process, 'exit').mockImplementation(() => {})
     const mockConsole = jest.spyOn(console, 'error').mockImplementation(() => {})
-    const unhandledRejection = require('../../../../app/exception/unhandledRejection')
+    const unhandledRejection = require('../../../../app/exception/unhandled-rejection')
     await unhandledRejection()
     expect(mockExit).toHaveBeenCalledTimes(1)
     expect(mockConsole).toHaveBeenCalledTimes(1)
