+++ test/unit/app/server/start-server.test.js
@@ -2,7 +2,7 @@ describe('Init Server Test', () => {
   test('Should start server', async () => {
     const server = require('../../../../app/server/server')
     jest.mock('../../../../app/server/server')
-    const startServer = require('../../../../app/server/init-server')
+    const startServer = require('../../../../app/server/start-server')
     await startServer()
     expect(server.start).toHaveBeenCalledTimes(1)
   })
